fix: comprehensive security and RFC compliance fixes (v1.0.1)
Security:
- HPACK integer decode overflow protection (M>=32 guard, addition overflow)
- HPACK pointer arithmetic overflow fix (parse_string bounds check)
- HPACK resize multiplication overflow guard on 32-bit
- Parser OOB reads: HEADERS+Priority short payload, PUSH_PROMISE<4 bytes
- Parser return value checks in all 10 dispatch functions

RFC 7540 compliance:
- DATA on stream 0 → PROTOCOL_ERROR (§6.1)
- RST_STREAM on idle stream → PROTOCOL_ERROR (§6.4)
- HEADERS state validation (§5.1)
- GOAWAY stream marking direction fix (§6.8)
- Send-side flow control enforcement (§6.9.1)
- MAX_CONCURRENT_STREAMS enforcement (§5.1.2)
- Remote MaxFrameSize for DATA sending
- CONTINUATION after PUSH_PROMISE stream ID fix (§6.10)
- Stream state machine: Recv PUSH_PROMISE Idle→ReservedRemote

Bug fixes:
- Empty DATA+END_STREAM handling (send and receive)
- ReadData actually consumes data (memory leak fix)
- flush_buffer preserves data on send failure
- send_goaway/send_rst_stream send preface first
- Stream send functions return actual results
- HPACK: compressor_destroy memory leak, rebuild_elems runtime check
- date.h: gmtime thread safety, strftime %G→%Y
- INT_MASK unsigned, macros do-while, assertion type mismatch
- add_slice(slice&&) uses std::move

Cleanup:
- Remove unused gRPC static table extension (entries 62-85)
- Fix typo readed_length→read_length
- Add CurrentStateTyped(), ScopedBufferedMode
- Add debug assert_same_thread() checks
- Add integration/protocol compliance test suites
- Update docs and version to 1.0.1
@

+++ include/http2/transport.h
@@ -483,7 +483,7 @@ struct ConnectionInfo {
     bool received_goaway;              /**< Whether GOAWAY has been received. */
     bool sent_goaway;                  /**< Whether GOAWAY has been sent. */
     bool draining;                     /**< Whether the connection is in drain phase. */
-    int32_t connection_window;         /**< Current connection-level send window. */
+    int64_t connection_window;         /**< Current connection-level send window. */
 };
 
 // ============================================================================
